+++ components/03-modules/range/range.config.js
@@ -2,6 +2,7 @@ module.exports = {
   status: 'wip',
   preview: '@docs-only-styles',
   context: {
+    attributes: '',
     class: '',
     filter: {
       wrapperClass: '',
@@ -25,8 +26,10 @@ module.exports = {
         class: 'range-filter__label',
         text: 'Minimum value'
       },
-      'field': {
+      field: {
         class: 'range-filter__field range-filter__field--lower',
+        id: 'range-filter-min',
+        name: 'range-filter-min',
         placeholder: '',
         value: '0'
       }
@@ -39,9 +42,13 @@ module.exports = {
       },
       field: {
         class: 'range-filter__field range-filter__field--upper',
+        id: 'range-filter-max',
+        name: 'range-filter-max',
         placeholder: '',
         value: '1000'
       }
-    }
-  }
+    },
+    script: true,
+    showFloat: 0
+  },
 };