+++ components/03-modules/range/range.js
@@ -10,59 +10,60 @@
   }
 
   function createRangeSlider(rangeFilter) {
-
-    const rangeSlider = rangeFilter.querySelector('.range-filter__slider'),
-          inputMax = rangeFilter.querySelector('.range-filter__field--upper'),
-          inputMin = rangeFilter.querySelector('.range-filter__field--lower');
+    const inputMax = rangeFilter.querySelector('.range-filter__field--upper'),
+          inputMin = rangeFilter.querySelector('.range-filter__field--lower'),
+          rangeSlider = rangeFilter.querySelector('.range-filter__slider');
 
     rangeSlider.classList.add('noUi-extended');
     initNoUiSlider(rangeSlider);
     bindEvents(rangeSlider, inputMin, inputMax);
-
   }
 
   function initNoUiSlider(rangeSlider) {
+    let minValue = parseInt(rangeSlider.getAttribute('data-min')),
+        maxValue = parseInt(rangeSlider.getAttribute('data-max'));
+
+    if (!minValue) {
+      minValue = 0;
+    }
+    if (!maxValue) {
+      maxValue = 1000;
+    }
 
     noUiSlider.create(rangeSlider, {
-      start: [0, 1000],
+      start: [minValue, maxValue],
       connect: true,
-      tooltips: [
-        true,
-        true
-      ],
+      step: 1,
       range: {
-        'min': 0,
-        'max': 200
-      },
-      formatter: {
-        to(value) {
-          return value + ',-';
-        },
-        from(value) {
-          return value.replace(',-', '');
-        }
-      },
+        'min': minValue,
+        'max': maxValue
+      }
     });
-
   }
 
   function bindEvents(rangeSlider, inputMin, inputMax) {
+    const showFloat = parseInt(rangeSlider.getAttribute('data-show-float'));
 
     rangeSlider.noUiSlider.on('update', (values) => {
-      inputMin.value = values[0];
-      inputMax.value = values[1];
+      if (showFloat) {
+        inputMin.value = values[0];
+        inputMax.value = values[1];
+      }
+      else {
+        inputMin.value = parseInt(values[0]);
+        inputMax.value = parseInt(values[1]);
+      }
     });
 
-    inputMax.addEventListener('change', () => {
-      rangeSlider.noUiSlider.set([null, this.value]);
+    inputMax.addEventListener('change', (event) => {
+      rangeSlider.noUiSlider.set([null, event.target.value]);
     });
 
-    inputMin.addEventListener('change', () => {
-      rangeSlider.noUiSlider.set([null, this.value]);
+    inputMin.addEventListener('change', (event) => {
+      rangeSlider.noUiSlider.set([null, event.target.value]);
     });
   }
 
   init();
-
 
 }());